moved open section into mobile view
