+++ packages/space-opera/src/components/model_viewer_snippet/model_viewer_snippet.ts
@@ -18,7 +18,7 @@
 import '@material/mwc-button';
 import './components/open_button.js';
 import './components/download_button.js';
-import './components/open_mobile_view.js';
+import '../mobile_view/open_mobile_view.js';
 import '../shared/snippet_viewer/snippet_viewer.js';
 import '../shared/expandable_content/expandable_tab.js';
